requirements.txt
gradio==5.15.0
numpy==2.1.1
opencv-python==4.11.0.86
torch==2.6.0
torchvision==0.21.0
ultralytics==8.3.78
Pillow==10.0.0